requirements.txt
fastapi
uvicorn[standard]
python-multipart
langchain
langgraph
langchain-groq
langchain-community
pypdf
faiss-cpu
numpy<2.0
sentence-transformers
pydantic-settings
pytest
httpx